Dataviews: Extract to dedicated bundled package
