+++ packages/edit-site/src/components/sidebar-dataviews/default-views.js
@@ -3,11 +3,7 @@
  */
 import { __ } from '@wordpress/i18n';
 import { trash } from '@wordpress/icons';
-
-/**
- * Internal dependencies
- */
-import { LAYOUT_TABLE, OPERATOR_IN } from '../dataviews';
+import { LAYOUT_TABLE, OPERATOR_IN } from '@wordpress/dataviews';
 
 const DEFAULT_PAGE_BASE = {
 	type: LAYOUT_TABLE,
