+++ packages/edit-site/src/style.scss
@@ -1,10 +1,10 @@
 @import "../../interface/src/style.scss";
+@import "../../dataviews/src/style.scss";
 
 @import "./components/add-new-template/style.scss";
 @import "./components/block-editor/style.scss";
 @import "./components/canvas-loader/style.scss";
 @import "./components/code-editor/style.scss";
-@import "./components/dataviews/style.scss";
 @import "./components/global-styles/style.scss";
 @import "./components/global-styles/screen-revisions/style.scss";
 @import "./components/header-edit-mode/style.scss";
