+++ tools/webpack/packages.js
@@ -29,6 +29,7 @@ const BUNDLED_PACKAGES = [
 	'@wordpress/interface',
 	'@wordpress/undo-manager',
 	'@wordpress/sync',
+	'@wordpress/dataviews',
 ];
 
 // PHP files in packages that have to be copied during build.
